complete esaboraManager test
