+++ tests/Unit/Service/Esabora/EsaboraSCHSServiceTest.php
@@ -8,7 +8,6 @@ use App\Service\Esabora\Response\DossierStateSCHSResponse;
 use App\Service\UploadHandlerService;
 use App\Tests\FileHelper;
 use App\Tests\FixturesHelper;
-use App\Tests\Unit\Messenger\DossierMessageTrait;
 use PHPUnit\Framework\MockObject\MockObject;
 use Psr\Log\LoggerInterface;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
@@ -19,7 +18,6 @@ use Symfony\Component\HttpFoundation\Response;
 
 class EsaboraSCHSServiceTest extends KernelTestCase
 {
-    use DossierMessageTrait;
     use FileHelper;
     use FixturesHelper;
